文件上传

+++ src/main/java/com/mettre/usually/service/FileService.java
@@ -7,7 +7,7 @@ public interface FileService {
 
     int deleteByPrimaryKey(Long fileId);
 
-    int insert(MultipartFile multipartFile);
+    long insert(MultipartFile multipartFile);
 
     int insertSelective(File record);
